Add an option to show all lookups
See https://github.com/simoncozens/crowbar/issues/24

Does not actually work, but I’m sending it just in case someone can make
a head or tail of it.

Clicking on the checkbox with cause an error on the console with a link
to https://reactjs.org/docs/forms.html#controlled-components which is
all Greek to me.

I also had to disable react/jsx-wrap-multilines lint rule, because the
linter can’t make its mind whether it wants parenthesis or not.